+++ package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "axios": "^1.6.8",
     "cheerio": "^1.0.0-rc.12",
+    "csv-writer": "^1.6.0",
     "express": "^4.17.1",
     "node-fetch": "^2.7.0"
   },
