+++ src/data/mobileData.csv
@@ -1,0 +1,25 @@
+NAME,PRICE,RATING,SPECS
+"realme P1 5G (Phoenix Red, 128 GB)","₹16,999",4.4,"8 GB RAM | 128 GB ROM | Expandable Upto 2 TB,16.94 cm (6.67 inch) Full HD+ Display,50MP + 2MP | 16MP Front Camera,5000 mAh Battery,Dimensity 7050 Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for Inbox Accessories"
+"SAMSUNG Galaxy F15 5G (Groovy Violet, 128 GB)","₹11,999",4.2,"4 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"OPPO Reno11 5G (Wave Green, 256 GB)","₹29,999",4.3,"8 GB RAM | 256 GB ROM | Expandable Upto 2 TB,17.02 cm (6.7 inch) Full HD+ Display,50MP + 8MP + 32MP | 32MP Front Camera,5000 mAh Battery,Mediatek Dimensity 7050 Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for Inbox Accessories"
+"REDMI 13C 5G (Startrail Silver, 256 GB)","₹13,999",4.2,"8 GB RAM | 256 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"REDMI 13C 5G (Starlight Black, 256 GB)","₹13,999",4.2,"8 GB RAM | 256 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"REDMI 13C 5G (Startrail Green, 256 GB)","₹13,999",4.2,"8 GB RAM | 256 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"SAMSUNG Galaxy F15 5G (Jazzy Green, 128 GB)","₹12,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"Tecno POVA 5 Pro 5G (Dark Illusion, 256 GB)","₹15,999",4.2,"8 GB RAM | 256 GB ROM | Expandable Upto 1 TB,17.22 cm (6.78 inch) Full HD+ Display,50MP + 0MP | 16MP Front Camera,5000 mAh Battery,MediaTek Dimensity 6080 Processor,Domestic Warranty of 12 Months on Phone & 6 Months on Accessories"
+"realme Narzo N53 (Feather Gold, 128 GB)","₹8,990",4.2,"8 GB RAM | 128 GB ROM,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera,5000 mAh Battery,Android 13, Realme UI T Processor,Domestic 1 Year on Handset and 6 Months on Accessories"
+"SAMSUNG Galaxy F54 5G (Stardust Silver, 256 GB)","₹22,999",4.2,"8 GB RAM | 256 GB ROM | Expandable Upto 1 TB,17.02 cm (6.7 inch) Full HD+ Display,108MP + 8MP + 2MP | 32MP Front Camera,6000 mAh Battery,Octa Core Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"realme P1 Pro 5G (Phoenix Red, 128 GB)","₹21,999",4.6,"8 GB RAM | 128 GB ROM,17.02 cm (6.7 inch) Full HD+ Display,50MP + 8MP | 16MP Front Camera,5000 mAh Battery,6 Gen 1 Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for Inbox Accessories"
+"SAMSUNG Galaxy F15 5G (Jazzy Green, 128 GB)","₹11,999",4.2,"4 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"SAMSUNG Galaxy F15 5G (Ash Black, 128 GB)","₹11,999",4.2,"4 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"REDMI 13C 5G (Starlight Black, 128 GB)","₹11,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"REDMI 13C 5G (Startrail Green, 128 GB)","₹11,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"REDMI 13C 5G (Startrail Silver, 128 GB)","₹11,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"Tecno Phantom V fold 5G Full Size Fold LTPO AMOLED Display (Black, 256 GB)","₹69,999",4.4,"12 GB RAM | 256 GB ROM,16.31 cm (6.42 inch) Display,50MP Rear Camera,5000 mAh Battery,12 month on smart phone & 6 months on accessories in the box domestic warranty"
+"Google Pixel 8 (Rose, 128 GB)","₹62,999",4.3,"8 GB RAM | 128 GB ROM,15.75 cm (6.2 inch) Full HD+ Display,50MP + 12MP | 10.5MP Front Camera,4575 mAh Battery,Tensor G3 Processor,1 Year Domestic Warranty"
+"Motorola g14 (Sky Blue, 128 GB)","₹8,499",4.1,"4 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 2MP | 8MP Front Camera,5000 mAh Battery,T616 Processor,Immersive Stereo Speakers with Dolby Atmos,1 Year on Handset and 6 Months on Accessories"
+"REDMI 13C 5G (Startrail Silver, 128 GB)","₹10,499",4.2,"4 GB RAM | 128 GB ROM | Expandable Upto 1 TB,17.12 cm (6.74 inch) HD+ Display,50MP Rear Camera | 5MP Front Camera,5000 mAh Battery,Mediatek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"SAMSUNG Galaxy F15 5G (Ash Black, 128 GB)","₹12,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"SAMSUNG Galaxy F15 5G (Groovy Violet, 128 GB)","₹12,999",4.2,"6 GB RAM | 128 GB ROM | Expandable Upto 1 TB,16.51 cm (6.5 inch) Full HD+ Display,50MP + 5MP + 2MP | 13MP Front Camera,6000 mAh Lithium ion Battery,MediaTek Dimensity 6100+ Processor,1 Year Manufacturer Warranty for Device and 6 Months Manufacturer Warranty for In-Box Accessories"
+"POCO X6 Pro 5G (Yellow, 512 GB)","₹25,999",4.3,"12 GB RAM | 512 GB ROM,16.94 cm (6.67 inch) Display,64MP + 8MP + 2MP | 16MP Front Camera,5000 mAh Battery,Dimensity D8300 Ultra Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
+"OPPO Find N3 Flip (Cream Gold, 256 GB)","₹89,990",4.4,"12 GB RAM | 256 GB ROM,17.27 cm (6.8 inch) Full HD Display,50MP + 48MP + 32MP | 32MP Front Camera,4300 mAh Battery,Mediatek Dimensity 9200 Processor,1 Year Manufacturer Warranty for Phone and 6 Months Warranty for In the Box Accessories"
