+++ src/data/tshirtData.csv
@@ -1,0 +1,21 @@
+NAME,PRICE
+UrbanMark Men 100% Cotton Half Sleeves Regular Fit Solid Polo T Shirt-Light Blue,Rs.  384
+UNIBERRY - Red Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  346
+ADORATE - Turquoise Cotton Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  314
+MXN - White Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  346
+MXN - White Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  354
+UrbanMark Mens Regular Fit Round Neck Full Sleeves Colorblock T Shirt -Beige & Grey,Rs.  384
+BULLMER - Mustard Cotton Blend Regular Fit Men's T-Shirt ( Pack of 1 ),Rs.  412
+Veirdo - Multicolor Cotton Oversized Fit Men's T-Shirt ( Pack of 1 ),Rs.  419
+BULLMER - Olive Cotton Blend Regular Fit Men's T-Shirt ( Pack of 1 ),Rs.  412
+BULLMER - Sea Green Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  437
+Veirdo - Purple Cotton Regular Fit Men's T-Shirt ( Pack of 1 ),Rs.  356
+AUSK Cotton Blend Regular Fit Solid Half Sleeves Men's T-Shirt - Black ( Pack of 1 ),Rs.  210
+Leotude - Grey Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  370
+Leotude - Black Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  370
+Leotude - Mustard Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  370
+The Million Club - Maroon Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  290
+BULLMER - Maroon Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  356
+BULLMER - Green Cotton Blend Regular Fit Men's Polo T Shirt ( Pack of 1 ),Rs.  356
+AUSK - Multicolor Cotton Regular Fit Men's T-Shirt ( Pack of 1 ),Rs.  250
+The Million Club - Light Blue Cotton Blend Regular Fit Men's T-Shirt ( Pack of 1 ),Rs.  234
