+++ src/routes/routes.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const { fetchFlipkartMobileProducts,
-    fetchSnapdealTShirtProducts,
-    fetchFullFlipkartMobileProducts} = require('../controllers/controllers.js') 
+const { fetchFlipkartMobileProducts} = require('../controllers/fetchFlipkartMobileProducts');
+const {fetchSnapdealTShirtProducts} = require('../controllers/fetchSnapdealTShirtProducts'); 
+const { fetchFullFlipkartMobileProducts} = require('../controllers/fetchFullFlipkartMobileProducts')
 
 router.get('/fetch/flipkart/mobile', fetchFlipkartMobileProducts);
 router.get('/fetch/snapdeal/t-shirt', fetchSnapdealTShirtProducts);
